Project/requirements.txt
google-cloud-pubsub==2.18.0
requests==2.26.0
Flask==2.0.1  # Add this line


